src/gpu/models/predictor.py
import torch
import torch.nn as nn
import torchvision.transforms as transforms
from PIL import Image
import timm
from transformers import ViTForImageClassification, ViTImageProcessor
import open_clip

import os
import random
import time

# --------------------------------------------------------------------------
# ★★ 중요 ★★
# 제공해주신 3개의 학습 노트북(DINO, ViT, CLIP) 내용을 기반으로
# 실제 모델 구조와 가중치 로딩 방식을 100% 동일하게 구현한 최종 버전입니다.
# --------------------------------------------------------------------------

# --- 모델 전체를 담는 Wrapper 클래스 정의 ---


# DINO 모델 정의 (Backbone + Classifier)
class DinoClassifier(nn.Module):
    def __init__(self, num_classes=2):
        super().__init__()
        # 노트북과 동일하게 timm에서 'vit_small_patch16_224_dino' 모델을 불러옵니다.
        self.backbone = timm.create_model(
            "vit_small_patch16_224_dino", pretrained=False
        )
        self.backbone.head = nn.Identity()  # 원본 head 제거

        # 별도의 Classifier를 정의합니다.
        self.classifier = nn.Linear(self.backbone.embed_dim, num_classes)

    def forward(self, x):
        features = self.backbone(x)
        return self.classifier(features)


# --- 설정 ---

# 모델 파일들이 저장된 경로
MODEL_PATH_DIR = "model_files"
MODEL_PATHS = {
    "DINO": os.path.join(MODEL_PATH_DIR, "DINO_v1.pth"),
    "ViT": os.path.join(MODEL_PATH_DIR, "ViT_v1.pth"),
    "CLIP": os.path.join(MODEL_PATH_DIR, "CLIP_v1.pth"),
}

# 로드된 모델, 이미지 프로세서를 저장할 딕셔너리
MODELS = {}
PROCESSORS = {}

# --- 모델 로드 함수 ---


def load_models():
    """서버 시작 시 모델과 프로세서를 미리 로드하여 메모리에 올려놓는 함수"""
    print("Loading models...")
    device = "cuda" if torch.cuda.is_available() else "cpu"
    print(f"Using device: {device}")

    # --- 1. DINO 모델 로드 ---
    dino_path = MODEL_PATHS["DINO"]
    if os.path.exists(dino_path):
        try:
            # 모델 인스턴스 생성
            model = DinoClassifier(num_classes=2)

            # 저장된 state_dict 불러오기 (backbone과 classifier가 분리된 딕셔너리 형태)
            state_dict = torch.load(dino_path, map_location=device)
            model.backbone.load_state_dict(state_dict["backbone"])
            model.classifier.load_state_dict(state_dict["classifier"])

            model.to(device)
            model.eval()
            MODELS["DINO"] = {"model": model, "device": device}

            # DINO용 이미지 전처리기
            PROCESSORS["DINO"] = transforms.Compose(
                [
                    transforms.Resize((224, 224)),
                    transforms.ToTensor(),
                    transforms.Normalize(
                        mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]
                    ),
                ]
            )
            print(f"✅ DINO model loaded successfully from {dino_path}")
        except Exception as e:
            print(f"❗️ Error loading DINO model: {e}")
            MODELS["DINO"] = None
    else:
        print(f"❗️ DINO model file not found at {dino_path}")
        MODELS["DINO"] = None

    # --- 2. ViT 모델 로드 ---
    vit_path = MODEL_PATHS["ViT"]
    if os.path.exists(vit_path):
        try:
            model_name = "google/vit-base-patch16-224-in21k"
            # ViT 모델 구조를 HuggingFace Hub에서 불러옵니다. (id2label, label2id는 추론에 필수)
            model = ViTForImageClassification.from_pretrained(
                model_name,
                num_labels=2,
                id2label={0: "REAL", 1: "FAKE"},
                label2id={"REAL": 0, "FAKE": 1},
            )
            # 저장된 가중치를 불러옵니다.
            model.load_state_dict(torch.load(vit_path, map_location=device))
            model.to(device)
            model.eval()
            MODELS["ViT"] = {"model": model, "device": device}

            # ViT용 이미지 전처리기
            PROCESSORS["ViT"] = ViTImageProcessor.from_pretrained(model_name)
            print(f"✅ ViT model loaded successfully from {vit_path}")
        except Exception as e:
            print(f"❗️ Error loading ViT model: {e}")
            MODELS["ViT"] = None
    else:
        print(f"❗️ ViT model file not found at {vit_path}")
        MODELS["ViT"] = None

    # --- 3. CLIP 모델 로드 ---
    clip_path = MODEL_PATHS["CLIP"]
    if os.path.exists(clip_path):
        try:
            # CLIP 모델 구조와 이미지 전처리기를 open_clip 라이브러리에서 불러옵니다.
            model, _, preprocess = open_clip.create_model_and_transforms(
                "ViT-B-32", pretrained="laion2b_s34b_b79k"
            )
            # 모델의 classification head 부분을 노트북과 동일하게 수정
            model.visual.head = nn.Linear(768, 2)

            # 저장된 가중치를 불러옵니다.
            model.load_state_dict(torch.load(clip_path, map_location=device))
            model.to(device)
            model.eval()
            MODELS["CLIP"] = {"model": model, "device": device}

            # CLIP용 이미지 전처리기
            PROCESSORS["CLIP"] = preprocess
            print(f"✅ CLIP model loaded successfully from {clip_path}")
        except Exception as e:
            print(f"❗️ Error loading CLIP model: {e}")
            MODELS["CLIP"] = None
    else:
        print(f"❗️ CLIP model file not found at {clip_path}")
        MODELS["CLIP"] = None


# --- 추론 함수 ---
def run_prediction(image_path, model_name):
    """이미지 경로와 모델 이름을 받아 딥페이크 여부를 판별합니다."""
    print(f"Running prediction with model: {model_name} on image: {image_path}")

    if model_name not in MODELS or MODELS.get(model_name) is None:
        # 모델이 로드되지 않았다면, 가짜 추론 결과 반환
        print(f"Warning: {model_name} model not loaded. Returning mock result.")
        time.sleep(random.uniform(0.5, 1.5))
        confidence = random.random()
        label = (
            "딥페이크" if confidence > 0.7 else ("실제" if confidence < 0.3 else "애매")
        )
        return {"label": label, "confidence": confidence}

    try:
        model_info = MODELS[model_name]
        model, device = model_info["model"], model_info["device"]
        processor = PROCESSORS[model_name]

        # 1. 이미지 로드 및 전처리
        image = Image.open(image_path).convert("RGB")

        # 모델별로 다른 전처리 방식 적용
        if model_name in ["DINO", "CLIP"]:
            image_tensor = processor(image).unsqueeze(0).to(device)
        elif model_name == "ViT":
            inputs = processor(images=image, return_tensors="pt").to(device)
            image_tensor = inputs["pixel_values"]

        # 2. 모델 추론
        with torch.no_grad():
            if model_name == "CLIP":
                # CLIP은 이미지 인코더 부분만 사용
                output = model.visual(image_tensor)
            else:  # DINO, ViT
                output = model(image_tensor)

            # ViT는 출력이 딕셔너리 형태일 수 있음
            if isinstance(output, dict):
                output = output.logits

            probabilities = torch.nn.functional.softmax(output, dim=1)

        # 3. 결과 해석 (0: REAL, 1: FAKE 로 통일)
        # 딥페이크(FAKE)일 확률을 confidence로 사용합니다.
        # 노트북에서 REAL=0, FAKE=1로 라벨링 한 것을 확인했습니다.
        confidence = probabilities[0][1].item()

        if confidence > 0.7:
            label = "딥페이크"
        elif confidence < 0.3:
            label = "실제"
        else:
            label = "애매"

        print(f"Prediction result: {label}, Confidence: {confidence:.2f}")
        return {"label": label, "confidence": confidence}

    except Exception as e:
        import traceback

        traceback.print_exc()  # 자세한 오류 출력을 위해 추가
        return {"error": f"추론 중 오류 발생: {e}"}
