requirements.gpu.txt
Flask
Pillow
APScheduler
torch
torchvision
timm
transformers
open-clip-torch

# --- 팀의 다른 모델에 필요한 라이브러리를 아래에 추가하세요 ---
# 예:
# transformers